requirements.txt
pyautogui
requests
pillow
numpy
mss
